+++ src/IIS.Tests/Cake.IIS.Tests.csproj
@@ -78,6 +78,7 @@
   <ItemGroup>
     <Compile Include="Tests\ApplicationPoolTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Tests\ApplicationTests.cs" />
     <Compile Include="Tests\WebFarmTests.cs" />
     <Compile Include="Utils\DebugLog.cs" />
     <Compile Include="Utils\CakeHelper.cs" />
